+++ src/signalingprotocol/SignalingProtocol.js
@@ -7754,6 +7754,8 @@ $root.SdkMetric = (function() {
             case 100:
             case 101:
             case 102:
+            case 103:
+            case 104:
                 break;
             }
         if (message.value != null && message.hasOwnProperty("value"))
@@ -8061,6 +8063,14 @@ $root.SdkMetric = (function() {
         case 102:
             message.type = 102;
             break;
+        case "VIDEO_CODEC_DEGRADATION_HIGH_ENCODE_CPU":
+        case 103:
+            message.type = 103;
+            break;
+        case "VIDEO_CODEC_DEGRADATION_HW_ENCODE_FAILURE":
+        case 104:
+            message.type = 104;
+            break;
         }
         if (object.value != null)
             message.value = Number(object.value);
@@ -8191,6 +8201,8 @@ $root.SdkMetric = (function() {
      * @property {number} RTC_SPK_TOTAL_RECOVERED_RED=100 RTC_SPK_TOTAL_RECOVERED_RED value
      * @property {number} RTC_SPK_TOTAL_RECOVERED_FEC=101 RTC_SPK_TOTAL_RECOVERED_FEC value
      * @property {number} VIDEO_QUALITY_LIMITATION_DURATION_CPU=102 VIDEO_QUALITY_LIMITATION_DURATION_CPU value
+     * @property {number} VIDEO_CODEC_DEGRADATION_HIGH_ENCODE_CPU=103 VIDEO_CODEC_DEGRADATION_HIGH_ENCODE_CPU value
+     * @property {number} VIDEO_CODEC_DEGRADATION_HW_ENCODE_FAILURE=104 VIDEO_CODEC_DEGRADATION_HW_ENCODE_FAILURE value
      */
     SdkMetric.Type = (function() {
         var valuesById = {}, values = Object.create(valuesById);
@@ -8264,6 +8276,8 @@ $root.SdkMetric = (function() {
         values[valuesById[100] = "RTC_SPK_TOTAL_RECOVERED_RED"] = 100;
         values[valuesById[101] = "RTC_SPK_TOTAL_RECOVERED_FEC"] = 101;
         values[valuesById[102] = "VIDEO_QUALITY_LIMITATION_DURATION_CPU"] = 102;
+        values[valuesById[103] = "VIDEO_CODEC_DEGRADATION_HIGH_ENCODE_CPU"] = 103;
+        values[valuesById[104] = "VIDEO_CODEC_DEGRADATION_HW_ENCODE_FAILURE"] = 104;
         return values;
     })();